+++ src/api.cc
@@ -5,7 +5,7 @@ using namespace ctypes;
 
 static_assert(sizeof(packedvalue_handle) == sizeof(PackedValue));
 
-int CTIRegistryListNames(const char* tag, OUT int* ret_size, OUT const char *** ret_names) {
+int CTIRegistryListNames(const char* tag, OUT size_t* ret_size, OUT const char*** ret_names) {
   static thread_local std::vector<std::string> names;
   static thread_local std::vector<const char*> names_p;
   CHECK_EQ(tag, Registry<PackedFunc>::RegistryName_);
@@ -26,7 +26,7 @@ int CTIRegistryGet(const char* tag, const char* name, OUT func_handle* ret_handl
   return CTI_SUCCESS;
 }
 
-int CTIPackedFuncCall(const void* handle, int num_args, const unsigned* type_codes, const packedvalue_handle* values,
+int CTIPackedFuncCall(const void* handle, size_t num_args, const unsigned* type_codes, const packedvalue_handle* values,
     OUT unsigned* ret_type, OUT packedvalue_handle* ret_val) {
   PackedFunc::FuncCall(handle, num_args, static_cast<const PackedType*>(type_codes), reinterpret_cast<const PackedValue*>(values),
       static_cast<PackedType*>(ret_type), reinterpret_cast<PackedValue*>(ret_val));
